fix formula luas Lingkaran

+++ com/base/BangunDatar.java
@@ -6,6 +6,6 @@ public interface BangunDatar{
     void tampilluas();
     int panjang();
     int lebar();
-    int luas();
+    double luas();
 
 }
